#003 - New function to verify header, new function to push changes and refactoring MoveScript
main: Modification to make the new method calls, changed the variable dest_path to proj_path, changing way of calling move_to method, new user input to verify pipeline execution, correction on logs and moving of files after confirmation of pipeline successful execution

MoveScript: refactoring separating the move_to method into script_destination that construct and verify the path that the scripts should go into the project letting the move_to only work on moving files

GitManager: adding push_changes method that push the changes to the remote repository

FileEditor: adding verify_header function to verify if after the execution on the pipeline the files have the header added by executing it

+++ Source/MoveScript.py
@@ -1,14 +1,10 @@
 import shutil
-import os
 from pathlib import Path
 from datetime import datetime
 
 class MoveScript:
     @staticmethod
-    def move_to(source_files, destination_path, team_name):
-
-        files = []
-
+    def script_destination(destination_path, team_name):
         year = datetime.now().strftime("%Y")
         month = datetime.now().strftime("%m")
         path_base = Path(destination_path) / "sql"
@@ -16,28 +12,45 @@ class MoveScript:
         path_final = path_team / year / month
 
         try:
-
             if not path_team.exists():
                 print(f"Error: Team directory '{team_name}' not found in {path_base}")
-                return[]
+                return None
 
             if not path_final.exists():
                 print(f"Creating directory structure: {datetime.now().strftime('%Y/%m')}")
                 path_final.mkdir(parents=True, exist_ok=True)
-
-            for file_path in source_files:
-                src = Path(file_path)
-
-                if src.exists():
-                    destination = path_final/src.name
-                    shutil.move(str(src), str(destination))
-                    files.append(str(destination))
-
-                else:
-                    print(f"Warning: File not found: {file_path}")
-
-            return files
+            
+            return path_final
         
         except Exception as e:
-            print(f"Unexpected error while moving files: {e}")
+            print(f"Unexpected error while : {e}")
+            return None
+
+
+    @staticmethod
+    def move_to(file_path, destination_path, task_id):
+
+        task_id_str = str(task_id)
+
+        try:
+            src = Path(file_path)
+            dest_dir = Path(destination_path)
+
+            if src.exists() and task_id_str in src.name:
+                destination = dest_dir/src.name
+
+                print(f"Moving: {src.name} -> {dest_dir}")
+                shutil.move(str(src), str(destination))
+                return str(destination)
+            
+            if not src.exists():
+                print(f"⚠️ Warning: File not found: {file_path}")
+
+            elif task_id_str not in src.name:
+                print(f"⏭️ Skipping: '{src.name}' (ID {task_id_str} not found in name)")
+            
+            return None
+        
+        except Exception as e:
+            print(f"Unexpected error while moving file {file_path}: {e}")
             return []
